modify

+++ Order-Processing-System/src/main/java/com/e_commerce/dto/ProductDto.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @Data
 public class ProductDto {
 	  private String name;
-	    private float price;
+	    private BigDecimal price;
 	    private String description;
 }
